+++ online-banking-system-backend/src/main/resources/application.properties
@@ -6,7 +6,7 @@ server.port=8080
 spring.datasource.url=jdbc:mysql://localhost:3306/banking?createDatabaseIfNotExist=true&useSSL=false&allowPublicKeyRetrieval=true
 spring.datasource.username=root
 
-spring.datasource.password=root
+spring.datasource.password=1234
 
 # JPA properties
 spring.jpa.show-sql = true
