+++ online-banking-system-frontend/src/App.jsx
@@ -1,18 +1,16 @@
 import "./App.css";
 import "bootstrap/dist/css/bootstrap.min.css";
-
 import Landing from "./components/Landing";
 import Login from "./components/Login";
 import { Route, Routes } from "react-router-dom";
 import Signup from "./components/Signup";
 import TransferFunds from "./components/TransferFunds";
-// import TransactionReport from "./components/TransactionReport";
 import Dashboard from "./components/Dashboard";
 import ConfirmPin from "./components/ConfirmPin";
 import DisplayAccounts from "./components/DisplayAccounts";
 import AccountHome from "./components/AccountHome";
-import Treport from "./components/Treport";
 import DepositMoney from "./components/DepositMoney";
+import TransactionReport from "./components/TransactionReport";
 
 function App() {
   return (
@@ -29,10 +27,8 @@ function App() {
         <Route index element={<AccountHome />} />
         <Route path="transferfunds" element={<TransferFunds />} />
         <Route path="depositmoney" element={<DepositMoney />} />
-        {/* <Route path="transactionslist" element={<TransactionReport />} /> */}
-        <Route path="transactionslist" element={<Treport />} />
+        <Route path="transactionslist" element={<TransactionReport />} />
       </Route>
-      {/* <Route path="/treport" element={<Treport/>}></Route> */}
     </Routes>
   );
 }
